% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_list_length.R
\name{add_list_length}
\alias{add_list_length}
\title{Add a context-specific list-length}
\usage{
add_list_length(
  df,
  taxa_col = "taxa",
  context,
  create_list_col = FALSE,
  sep = "_"
)
}
\arguments{
\item{df}{Data frame with taxa and context columns.}

\item{taxa_col}{Character column name in \code{df} with taxa.}

\item{context}{Character vector of columns names defining the context.}

\item{create_list_col}{Create a single column named \code{list} with the full
context separated by \code{sep}.}

\item{sep}{Separator between contexts if \code{create_list_col}}
}
\value{
Data frame with extra column(s) list length, named \code{sr}, and, optionally,
list name, \code{list}.
}
\description{
Optionally creating a single column \code{list} with the full context.
}
