% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ap}
\alias{ap}
\title{Stanreg model for \emph{Acacia paradoxa}}
\format{
object of class stanreg
glm
lm
gamm4
}
\usage{
ap
}
\description{
Stanreg model for \emph{Acacia paradoxa}
}
\keyword{datasets}
