% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_plot.R
\name{cat_plot}
\alias{cat_plot}
\title{Plot with increasing credible range error bars decreasing in size around
credible intervals}
\usage{
cat_plot(
  df,
  orig_data = NULL,
  y = "prop",
  x = "year",
  facet_col = "LSA",
  y_name = "reporting rate",
  likelihood_col = "pos_raw"
)
}
\arguments{
\item{df}{Dataframe with \code{y}, \code{y_prop} etc}

\item{orig_data}{Dataframe with original data}

\item{x, y}{Chracter name of columns in \code{df} and \code{orig_data} with response
variable}

\item{facet_col}{Character name of column in \code{df} with a column to facet on.}

\item{y_name}{Character. Name to give to \code{y} in plot}

\item{likelihood_col}{Character name of column in \code{df} with likelihoods}
}
\value{
\code{ggplot}
}
\description{
Option to include original data. Currently needs columns \code{q01}, \code{q05}, \code{q25},
\code{q50}, \code{q75}, \code{q95} and \code{q99} in \code{df}.
}
