% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_list_df.R
\name{create_list_df}
\alias{create_list_df}
\title{Filter data to a set of specifications}
\usage{
create_list_df(
  df,
  taxa_col = "taxa",
  geo_level,
  tax_level,
  minListLengthThresh = 3,
  maxListLengthOccurenceThresh = 3,
  minlistOccurenceThresh = 5,
  minYearsThresh = 3,
  minListLengthsThresh = 2,
  minYearSpanThresh = 20
)
}
\arguments{
\item{df}{Cleaned biological data.}

\item{taxa_col}{Character name of column in \code{df} with taxa.}

\item{geo_level}{Character name of column in \code{df} with geographic scale(s) of
interest}

\item{tax_level}{Character name of column(s) in \code{df} with taxonomic scale(s)
of interest}

\item{minListLengthThresh}{Minimum list length allowed}

\item{maxListLengthOccurenceThresh}{Maximum}

\item{minYearSpanThresh}{}
}
\value{

}
\description{
Currently this feels messy, much room for improvement.
}
