% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mod.R
\name{explore_mod}
\alias{explore_mod}
\title{Explore (pre- and most-) model results.}
\usage{
explore_mod(
  taxa,
  common,
  df,
  mod_path,
  mod_type,
  resp_var = "prop",
  exp_var,
  max_levels = 30,
  draws = 200,
  post_groups,
  tests = NULL,
  re_run = FALSE
)
}
\arguments{
\item{taxa}{Scientific name of taxa (for labelling things).}

\item{common}{Common name of taxa (also used in labels).}

\item{df}{Data used for model saved at \code{mod_path}.}

\item{mod_path}{Path to saved model.}

\item{mod_type}{Type of model (e.g. 'reporting rate', or 'occupancy')}

\item{resp_var}{What is the response variable?}

\item{exp_var}{What are the variables to include in model exploration?}

\item{max_levels}{Maximum number of classes to include in categorical plots.}

\item{draws}{Number of draws from posterior distribution to display in plots.}

\item{post_groups}{Character names of variables to include in model results.}

\item{tests}{Years at which to predict (and compare change) as dataframe
with type = c("reference", "recent") and year = four-digit year corresponding
to references and recent.}

\item{re_run}{Logical. If file \code{gsub("\\\\.rds", "_res.rds", mod_path)} already
exists, should it be re-run?}
}
\value{
A list with components
\itemize{
\item{count_char}{ggplot object. Count of levels within character variables.}
\item{y_vs_char}{ggplot object. Response variable vs character variables.}
\item{count_num}{ggplot object. Histogram of each numeric variable.}
\item{y_vs_num}{ggplot object. Response variable vs numeric variables with \code{geom_smooth()}.}
\item{pairs}{\code{GGally::ggpairs} object from \code{df}.}
\item{pred}{Dataframe of \code{posterior_predict} for levels of interest.}
\item{res}{\code{pred} summarised by \code{post_groups}.}
}
}
\description{
NOTE. Code still contains references to \code{geo2}.
}
