% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mod_II.R
\name{explore_mod_ll}
\alias{explore_mod_ll}
\title{Explore (pre-) and (post-) results from \code{make_mod_res} after \code{make_ll_model}}
\usage{
explore_mod_ll(
  model_results,
  plot_title = NULL,
  max_levels = 15,
  recent = as.numeric(format(Sys.Date(), "\%Y")) - 2,
  limit_preds = TRUE,
  plot_draws = 500,
  diff_direction = c("pos", "neg"),
  do_gc = TRUE
)
}
\arguments{
\item{model_results}{List. results from \code{make_ll_model()}}

\item{plot_title}{Character. Title used on all plots.}

\item{do_gc}{}
}
\value{

}
\description{
General workflow is make the model (\code{make_ll_model}), make the model results
(\code{make_mod_res}) then explore results (\code{explore_mod_ll}). Might work for
other rstanarm models? Not sure anymore.
}
