% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_list_df.R
\name{filter_list_df}
\alias{filter_list_df}
\title{Filter occurrence data to a set of specifications}
\usage{
filter_list_df(
  df,
  taxa_col = "taxa",
  min_length = 3,
  shortest_max = 3,
  min_occurrences = 5,
  min_years = 3,
  min_year = as.numeric(format(Sys.Date(), "\%Y")) - 11,
  max_year = as.numeric(format(Sys.Date(), "\%Y")) - 1
)
}
\arguments{
\item{df}{Cleaned biological data.}

\item{taxa_col}{Character name of column in \code{df} with taxa.}

\item{shortest_max}{Minimum allowable maximum length.}

\item{min_occurrences}{Minimum allowable occurrence.}

\item{min_years}{Minimum allowable years with occurrence.}

\item{min_year}{Must be records less than \code{min_year}.}

\item{max_year}{Much be records greater than \code{max_year}.}
}
\value{
Filtered data frame.
}
\description{
Works on results from \code{make_list_df}. Reworked to focus on only removing taxa
(i.e. not visits or other contexts).
}
