% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_mean.R
\name{geo_mean}
\alias{geo_mean}
\title{Geometric mean}
\usage{
geo_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric}

\item{na.rm}{Remove any \code{NA} values prior to calculation}
}
\value{

}
\description{
Defined as:
\loadmathjax
\mjdeqn{ \left(\prod_{i=1}^n x_i\right)^\frac{1}{n} = \sqrt[n]{x_1 x_2 \cdots x_n}
}{ASCII representation}
}
\details{
Based heavily on the \code{geoMean()} function in the \href{https://alexkowa.github.io/EnvStats/}{EnvStats}
package by \href{https://github.com/alexkowa}{alexkowa}, and various
\href{https://stackoverflow.com/questions/2602583/geometric-mean-is-there-a-built-in}{stackoverflow posts}.
}
\examples{
x <- rnorm(100, 1, 0.3)
geo_mean(x)

x <- c(x, NA)
geo_mean(x)
geo_mean(x, na.rm = TRUE)

}
