% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lulikelihood}
\alias{lulikelihood}
\title{Dataframe of likelihood thresholds and definitions}
\format{
\subsection{A data frame with 9 rows and
7 variables:
\describe{
\item{likelihood}{Factor. Definitions.}
\item{maxVal}{Double. Maximum proportion to fall in likelihood category}
\item{range}{Factor. Result of \code{cut(maxVal, breaks = c(0,.$maxVal))}}
\item{loose}{Factor wit+
0
-)` levels}
\item{very}{Factor wit++
+
0}{
}

\subsection{--)` levels}
\item{extreme}{Factor wit+++
++
+
0}{
}

\subsection{--
---)` levels}
\item{exceptional}{Factor wit++++
+++
++
+
0}{
}

\subsection{--}{

----)` levels}
...
}
}
}
\source{
\url{https://www.ipcc.ch/site/assets/uploads/2017/08/AR5_Uncertainty_Guidance_Note.pdf}
}
\usage{
lulikelihood
}
\description{
Following the intergovernmental panel on climate change \href{https://www.ipcc.ch/site/assets/uploads/2017/08/AR5_Uncertainty_Guidance_Note.pdf}{IPCC}.
}
\keyword{datasets}
