% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cooccur.R
\name{make_cooccur}
\alias{make_cooccur}
\title{Find taxa pairwise co-occurrence patterns}
\usage{
make_cooccur(pres_df, taxa_col = "taxa", context, ...)
}
\arguments{
\item{pres_df}{Dataframe with presence records, along with \code{taxa_col} and
\code{context}(s)}

\item{taxa_col}{Character name of columns in \code{pres_df} with taxa.}

\item{context}{Character name of column(s) in \code{pres_df} that define the
context.'}

\item{...}{Passed to \code{cooccur::cooccur()}}
}
\value{
a list of class \code{cooccur}
}
\description{
Find taxa pairwise co-occurrence patterns
}
