% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rr_model.R
\name{rr}
\alias{rr}
\title{Make, and save to disk, a reporting rate model.}
\usage{
rr(taxa, df, geo_cols, cell_cols = c("grid_l", "grid_s"), out_path, ...)
}
\arguments{
\item{taxa}{Character name of taxa for which model is being run. Used to name
output file.}

\item{df}{Cleaned, filtered data frame.}

\item{geo_cols}{Character name of columns in \code{df} containing geographic
context. Last index in \code{geo_cols} is used as primary analysis level.}

\item{cell_cols}{Character name of columns in \code{df} containing two scales of
(probably raster) grid cells. The first index is the larger cell, second
index is smaller.  The larger cell is used as a random factor in the model,
if there is more than one level.}

\item{out_path}{Path into which model results are saved.}

\item{...}{Passed to rstanarm::stan_gamm4 (e.g. chains, iter)}
}
\value{
\code{fs::path(out_path,paste0("reporting-rate_mod_",taxa,".rds"))}
}
\description{
Make, and save to disk, a reporting rate model.
}
