% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mod.R
\name{explore_mod}
\alias{explore_mod}
\title{Explore (pre- and post-) model results.}
\usage{
explore_mod(
  taxa,
  common,
  mod_path,
  out_file = gsub("\\\\.rds", "_summary.rds", mod_path),
  mod_type,
  resp_var = "prop",
  geo_var = "IBRA_SUB_N",
  time_var = "year",
  max_levels = 30,
  draws = 200,
  reference_year = 2000,
  recent_year = 2010,
  re_run = FALSE,
  quant_probs = c(0.05, 0.5, 0.95),
  binom_denom = "trials"
)
}
\arguments{
\item{taxa}{Scientific name of taxa (for labelling things).}

\item{common}{Common name of taxa (also used in labels).}

\item{mod_path}{Path to saved model.}

\item{out_file}{Path to save exploration results back to.}

\item{mod_type}{Type of model (e.g. 'reporting rate', or 'occupancy')}

\item{resp_var}{What is the response variable?}

\item{geo_var}{Categorical variable in the model (usually biogeographic)}

\item{time_var}{Continuous time variable (usually \code{year})}

\item{max_levels}{Maximum number of classes to include in categorical plots.}

\item{draws}{Number of draws from posterior distribution to display in plots.}

\item{reference_year}{Reference time at which to predict (and compare change)}

\item{recent_year}{Time at which to predict to compare against reference_year}

\item{re_run}{Logical. If file \code{out_file} already exists, should it be
re-run?}

\item{quant_probs}{Quantiles for summarising.}
}
\value{
A list with components
\describe{
\item{count_char}{ggplot object. Count of levels within character variables.}
\item{y_vs_char}{ggplot object. Response variable vs character variables.}
\item{count_num}{ggplot object. Histogram of each numeric variable.}
\item{y_vs_num}{ggplot object. Response variable vs numeric variables with \code{geom_smooth()}.}
\item{pairs}{\code{GGally::ggpairs} object from \code{df}.}
\item{pred}{Dataframe of \code{posterior_predict} for levels of interest.}
\item{res}{\code{pred} summarised by \code{c(resp_var, time_var)}.}
}
}
\description{
Explore (pre- and post-) model results.
}
