% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mod_II.R
\name{explore_mod_ll}
\alias{explore_mod_ll}
\title{Explore (pre-) and (post-) results from \code{make_mod_res} after \code{make_ll_model}}
\usage{
explore_mod_ll(
  model_results,
  taxa,
  common,
  max_levels = 15,
  recent = as.numeric(format(Sys.Date(), "\%Y")) - 2,
  limit_preds = TRUE,
  plot_draws = 200,
  diff_category = NULL,
  diff_direction = c("lower", "higher"),
  do_gc = TRUE,
  plot_divergent = FALSE,
  divergent_thresh = 0.001
)
}
\arguments{
\item{model_results}{List. results from \code{make_ll_model()}}

\item{plot_divergent}{Logical. If there are divergent transitions, generate
pairs plots? Can be very slow with many samples.}

\item{divergent_thresh}{Proportion of divergent transitions above which model
results will not be generated. Really, this should be 0. Default 0.001.}
}
\value{

}
\description{
General workflow is make the model (\code{make_ll_model}), make the model results
(\code{make_mod_res}) then explore results (\code{explore_mod_ll}). Might work for
other rstanarm models? Not sure anymore.
}
