% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_mod_II.R
\name{explore_mod_ll}
\alias{explore_mod_ll}
\title{Explore (pre-) and (post-) results from \code{make_mod_res} after \code{make_ll_model}}
\usage{
explore_mod_ll(
  model_results,
  plot_title = NULL,
  max_levels = 15,
  recent = as.numeric(format(Sys.Date(), "\%Y")) - 2,
  limit_preds = TRUE,
  plot_draws = 500,
  diff_direction = c("positive", "negative"),
  diff_category = NULL,
  facet_col = cat_col,
  dot_col = cov_col,
  do_gc = TRUE
)
}
\arguments{
\item{model_results}{List. results from \code{make_ll_model()}}

\item{plot_title}{Character. Title used on all plots.}

\item{diff_category}{Character name of column in original data to condition
on. If left as \code{NULL} will default to \code{cat_col} if there is more than one
level, or new category \code{all} if there in not. Useful to condition on
\code{random_col} if desired.}

\item{dot_col}{Character name of column in original data to colour the
original data points by in model results plots. Default is the secondary \code{x}
variable \code{cov_col}.}

\item{do_gc}{}
}
\description{
General workflow is make the model (\code{make_ll_model}), make the model results
(\code{make_mod_res}) then explore results (\code{explore_mod_ll}). Might work for
other rstanarm models? Not sure anymore.
}
