% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_list_df.R
\name{filter_list_df}
\alias{filter_list_df}
\title{Filter occurrence data to a set of specifications}
\usage{
filter_list_df(
  df,
  var_cols = "taxa",
  time_col = "year",
  shortest_max = 3,
  min_occurrences = 5,
  min_years = 3,
  min_span = 10,
  max_year = as.numeric(format(Sys.Date(), "\%Y")) - 1
)
}
\arguments{
\item{df}{Cleaned biological data.}

\item{var_cols}{Character name of column(s) in \code{df} that will be variables
in the model (or will define the model; e.g. \code{taxa_col})}

\item{time_col}{Character name of column in \code{df} with time variable. Usually
\code{year}.}

\item{shortest_max}{Minimum allowable maximum length.}

\item{min_occurrences}{Minimum allowable occurrence.}

\item{min_years}{Minimum allowable years with occurrence.}

\item{max_year}{Much be records greater than \code{max_year}.}

\item{min_year}{Must be records less than \code{min_year}.}
}
\value{
Filtered data frame.
}
\description{
Works on results from \code{make_list_df}. Probably works on other dfs.
}
