% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_list_df.R
\name{filter_list_df}
\alias{filter_list_df}
\title{Filter occurrence data to a set of specifications}
\usage{
filter_list_df(
  df,
  taxa_col = "taxa",
  geo_levels,
  tax_levels,
  time_levels,
  min_length = 3,
  shortest_max = 3,
  min_occurrence = 5,
  min_years = 3,
  min_lengths = 2,
  min_span = 20
)
}
\arguments{
\item{df}{Cleaned biological data.}

\item{taxa_col}{Character name of column in \code{df} with taxa.}

\item{min_length}{Minimum allowable list length.}

\item{shortest_max}{Minimum allowable maximum length.}

\item{min_occurrence}{Minimum allowable occurrence.}

\item{min_years}{Minimum allowable years with occurrence.}

\item{min_lengths}{Minimum allowable unique list lengths.}

\item{min_span}{Minimum allowable span of years.}

\item{geo_level}{Character name of column in \code{df} with geographic scale(s) of
interest}

\item{tax_level}{Character name of column(s) in \code{df} with taxonomic scale(s)
of interest}
}
\value{
Filtered data frame.
}
\description{
Currently this is clunky, much room for improvement.
}
