% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ll_model.R
\name{make_ll_model}
\alias{make_ll_model}
\title{Make, and save to disk, a list-length corrected reporting rate model.}
\usage{
make_ll_model(
  df,
  out_file,
  time_col = "year",
  geo_col = "geo",
  random_col = NULL,
  mod_type = "glm",
  k = if (mod_type == "gam") 6 else NULL,
  ...
)
}
\arguments{
\item{df}{Cleaned, filtered data frame.}

\item{out_file}{Path to file where model results are/will be saved.}

\item{geo_col}{Character name of columns in \code{df} containing geographic
context.}

\item{random_col}{Character name of column in \code{df} containing random factor
for model. This is usually the larger of two (probably raster) grid cell
sizes. Default is NULL, so no random factor in model.}

\item{mod_type}{Character. Currently "glm" or "gam".}

\item{k}{Integer used as \code{k} argument of \code{mgcv::s} for use in gam.}

\item{...}{Passed to \code{rstanarm::stan_gamm4} (e.g. chains, iter)}
}
\value{
\code{out_file}. Named list containing inputs and model results as \code{mod}.
}
\description{
Make, and save to disk, a list-length corrected reporting rate model.
}
