% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ll_model.R
\name{make_ll_model}
\alias{make_ll_model}
\title{Make, and save to disk, a list-length corrected reporting rate model.}
\usage{
make_ll_model(
  df,
  mod_file,
  var_col = "year",
  cat_col = "geo",
  cov_col = "list_length",
  random_col = NULL,
  log_cov = TRUE,
  rand_slope = TRUE,
  mod_type = "glm",
  k = if (mod_type == "gam") 6 else NULL,
  ...
)
}
\arguments{
\item{df}{Cleaned, filtered data frame.}

\item{var_col}{Chraacter name of column in \code{df} containing main \code{x} variable.}

\item{cat_col}{Character name of column in \code{df} containing categorical data.}

\item{cov_col}{Character name of column in \code{df} containing secondary \code{x}
variable.}

\item{random_col}{Character name of column in \code{df} containing random factor
for model. Default is NULL, so no random factor in model.}

\item{log_cov}{Logical. Use \code{cov_col} (\code{T}) or the log of \code{cov_col} (\code{F}) in
model specification.}

\item{rand_slope}{Logical. Should a random slope effect be included in the
model as well as any random intercept?}

\item{mod_type}{Character. Currently "glm" or "gam".}

\item{k}{Integer used as \code{k} argument of \code{mgcv::s} for use in gam.}

\item{...}{Passed to \code{rstanarm::stan_gamm4} (e.g. chains, iter)}

\item{out_file}{Path to file where model results are/will be saved.}
}
\value{
\code{out_file}. Named list containing inputs and model results as \code{mod}.
}
\description{
Make, and save to disk, a list-length corrected reporting rate model.
}
