% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mod_res.R
\name{make_mod_res}
\alias{make_mod_res}
\title{Summarise results from \code{make_ll_model}}
\usage{
make_mod_res(
  mod_file,
  ref = -20,
  pred_step = 2,
  include_random = FALSE,
  cov_q = NULL,
  cov_val = NULL,
  res_q = c(0.1, 0.5, 0.9),
  ps_thresh = 0.05,
  keep_mod = TRUE,
  keep_pred = TRUE,
  do_gc = TRUE,
  ...
)
}
\arguments{
\item{mod_file}{Character. Path to saved results from \code{make_ll_model}}

\item{ref}{Numeric. Either a single value of \code{var_col} from \code{make_ll_model}
to use as a reference value or a negative integer on the scale of \code{var_col}
to compare all values of \code{var_col} against \code{var_col + ref}.}

\item{pred_step}{Numeric. What step (in units of \code{var_col}) to predict at?}

\item{cov_q}{Numeric. What quantiles of \code{cov} to predict at?}

\item{cov_val}{Numeric. Alternative to cov_q. What value of \code{cov} to
predict at?}

\item{res_q}{Numeric. What quantiles to summarise predictions at?}

\item{ps_thresh}{Numeric value used in the \code{threshold} argument to
\code{bayestestR::p_significance()}. Set to zero to return probability of direction}

\item{keep_mod}{Logical. Return \code{mod} component of results (imported from
\code{mod_file}). Saves memory if this is not returned.}

\item{keep_pred}{Logical. Return \code{pred} component of results? Saves memory
if this is not returned.}

\item{do_gc}{Logical. Run \code{base::gc} after predict? On a server with shared
resources, can be necessary when summarising many, many models to prevent
RAM issues.}

\item{...}{Passed to \code{tidybayes::add_epred_draws()}}
}
\value{
A list with components
\describe{
\item{list_length}{data frame of list length quantiles in original data}
\item{data}{data frame of original data retrieved from \code{mod$data}}
\item{pred}{data frame of predictions (via
\code{tidybayes::add_epred_draws()}), including a \code{diff} column of the
difference between \code{ref}erence and each \code{pred_step} from the minimum of
\code{var_col} to the maximum of \code{var_col}}
\item{res}{data frame of predictions summarised, including columns for
each of \code{res_q} (applied to the differences between \code{ref} and \code{rec})}
\item{n_data}{\code{nrow(mod$data)}}
\item{n_fixed_coefs}{\verb{length(mod$coefficients[!grepl("b\\\[", names(mod$coefficients))])}}
\item{n_random_coefs}{\verb{length(mod$coefficients[grepl("b\\\[", names(mod$coefficients))])}}
}
}
\description{
Should now be generic enough to
}
