% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mod_res.R
\name{make_mod_res}
\alias{make_mod_res}
\title{Summarise results from \code{make_ll_model}}
\usage{
make_mod_res(
  path_to_model_file,
  scale_name,
  time_col = "year",
  random_col = NULL,
  ref = 2000,
  pred_step = 2,
  draws = 200,
  list_length_q = c(0.25, 0.5, 0.75),
  res_q = c(0.1, 0.5, 0.9),
  do_gc = TRUE,
  sample_new_levels_type = "uncertainty"
)
}
\arguments{
\item{path_to_model_file}{Character. Path to saved results from
\code{make_ll_model}}

\item{scale_name}{Character. Name of column in original data with spatial
data.}

\item{time_col}{Character. Name of column in original data with time data.}

\item{ref}{Numeric. Either a reference year (or other time point) or a
negative integer. If the former, all years will be compared to that
\code{ref}erence. If the later, all years will be compared to themselves + \code{ref}.}

\item{pred_step}{Numeric. What step (in units of \code{time_col}) to predict at?}

\item{draws}{Passed to \code{ndraws} argument of \code{tidybayes::add_epred_draws}.
Numeric or "max". Max will use all possible draws.}

\item{list_length_q}{Numeric. What list lengths quantiles to predict at?}

\item{res_q}{Numeric. What quantiles to summarise predictions at?}

\item{do_gc}{Logical. Run \code{base::gc} after predict? On a server with shared}

\item{sample_new_levels_type}{Character. Value of the \code{sample_new_levels}
argument to \code{tidybayes::add_epred_draws()}.
resources, can be necessary when summarising many, many models to prevent
filling RAM.}
}
\value{
A list with components
\describe{
\item{list_length}{data frame of list length quantiles in original data}
\item{data}{data frame of original data retrieved from \code{mod$data}}
\item{pred}{data frame of predictions (via
\code{tidybayes::add_epred_draws}), including a \code{diff} column of the
difference between \code{ref}erence and each \code{pred_step} from the minimum of
\code{time_col} to the maximum of \code{time_col}}
\item{res}{data frame of predictions summarised, including columns for
each of \code{res_q} (applied to the differences between \code{ref} and \code{rec})}
\item{n_data}{\code{nrow(mod$data)}}
\item{n_fixed_coefs}{\verb{length(mod$coefficients[!grepl("b\\\[", names(mod$coefficients))])}}
\item{n_random_coefs}{\verb{length(mod$coefficients[grepl("b\\\[", names(mod$coefficients))])}}
}
}
\description{
Summarise results from \code{make_ll_model}
}
