% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_overall.R
\name{make_overall}
\alias{make_overall}
\title{Combine results from other models into a single overall result}
\usage{
make_overall(taxa, common, year_diff_df)
}
\arguments{
\item{taxa}{Character. Name of taxa.}

\item{common}{Character. Common name of taxa.}

\item{year_diff_df}{Dataframe of differences between a reference and 'recent'
year.}
}
\value{
List with components
\itemize{
\item{overall}{Dataframe of summarised difference}
\item{overall_text}{A paragraph with rmarkdown formating describing the
result}
\item{overall_plot}{ggplot object with distribution of overall credible
values}
}
}
\description{
Combine results from other models into a single overall result
}
